requirements.txt
selenium==4.24.0
pytest==8.3.2
pytest-xdist==3.6.1
pytest-html==4.1.1
python-dotenv==1.0.1
webdriver-manager==4.0.2